@@ -1,49 +1,52 @@
 import 'mocha';
-import {expect} from 'chai';
+import { expect } from 'chai';
 import fetch from 'node-fetch';
 
 import * as loglevel from 'loglevel';
-import LogSelector from '../src/utils/LoggerSelector';
+import LogSelector from '../../../src/utils/LoggerSelector';
 LogSelector.setLogger(loglevel.getLogger('conseiljs'));
 LogSelector.setLevel('debug');
 
-import FetchSelector from '../src/utils/FetchSelector';
+import FetchSelector from '../../../src/utils/FetchSelector';
 FetchSelector.setFetch(fetch);
 
-import {ConseilQueryBuilder} from "../src/reporting/ConseilQueryBuilder";
-import {ConseilOperator, ConseilSortDirection} from "../src/types/conseil/QueryTypes"
-import {TezosConseilClient} from '../src/reporting/tezos/TezosConseilClient'
-import {OperationKindType} from "../src/types/tezos/TezosChainTypes";
-import {servers} from "./servers";
-
-const util = require('util');
-
-const ConseilV2URL = servers.conseilServer;
-const ConseilV2APIKey = servers.conseilApiKey;
+import { ConseilQueryBuilder } from '../../../src/reporting/ConseilQueryBuilder';
+import { ConseilOperator, ConseilSortDirection } from '../../../src/types/conseil/QueryTypes'
+import { TezosConseilClient } from '../../../src/reporting/tezos/TezosConseilClient'
+import { OperationKindType } from '../../../src/types/tezos/TezosChainTypes';
+import { conseilServer, faucetAccount } from '../../TestAssets.zeronet';
 
 describe('Tezos date interface test suite', () => {
     it('retrieve top block', async () => {
-        const result = await TezosConseilClient.getBlockHead({url: ConseilV2URL, apiKey: ConseilV2APIKey}, 'alphanet');
+        const result = await TezosConseilClient.getBlockHead(conseilServer, 'mainnet');
 
         expect(result['level']).to.be.greaterThan(625266, 'this may vary as the network changes');
         expect(result['baker'].length).to.be.greaterThan(0)
     });
 
     it('retrieve a block by hash', async () => {
-        const result = await TezosConseilClient.getBlock({url: ConseilV2URL, apiKey: ConseilV2APIKey}, 'alphanet', 'BKnMKWCeJwRtetQwuY5HRtbrsXPLyACFrygdwnM8jxAYcYEVkdd');
+        const result = await TezosConseilClient.getBlock(conseilServer, 'mainnet', 'BKnMKWCeJwRtetQwuY5HRtbrsXPLyACFrygdwnM8jxAYcYEVkdd');
 
         expect(result.length).to.equal(1);
     });
 
-    it('retrieve an account by address', async () => {
-        const result = await TezosConseilClient.getAccount({url: ConseilV2URL, apiKey: ConseilV2APIKey}, 'alphanet', 'tz1UzDGcT9JAZnryiDJmvGnNRMg3WzVnjLrn');
+    it('retrieve an operation group by hash', async () => {
+        const result = await TezosConseilClient.getOperationGroup(conseilServer, 'mainnet', 'ooH1GMyC7zHRP7SJgqnjzoxs5DshEUrYNehAQm9j5PUxMA4TdeP');
 
         expect(result.length).to.equal(1);
     });
-
-    //TODO: getOperationGroup
 
     //TODO: getOperationGroups
+
+    it('retrieve an operation group by hash', async () => {
+        let query = ConseilQueryBuilder.blankQuery();
+        query = ConseilQueryBuilder.addPredicate(query, 'operation_group_hash', ConseilOperator.EQ, ['ooH1GMyC7zHRP7SJgqnjzoxs5DshEUrYNehAQm9j5PUxMA4TdeP'], false);
+        query = ConseilQueryBuilder.setLimit(query, 10);
+
+        const result = await TezosConseilClient.getOperations(conseilServer, 'mainnet', query);
+
+        expect(result.length).to.equal(1);
+    });
 
     it('retrieve some blocks', async () => {
         let q = ConseilQueryBuilder.blankQuery();
@@ -51,15 +54,15 @@ describe('Tezos date interface test suite', () => {
         q = ConseilQueryBuilder.addOrdering(q, 'level', ConseilSortDirection.ASC);
         q = ConseilQueryBuilder.setLimit(q, 10);
 
-        const result = await TezosConseilClient.getBlocks({url: ConseilV2URL, apiKey: ConseilV2APIKey}, 'alphanet', q);
+        const result = await TezosConseilClient.getBlocks(conseilServer, 'mainnet', q);
 
         expect(result.length).to.equal(10);
-        expect(parseInt(result[9]["level"])).to.greaterThan(parseInt(result[1]["level"]));
-        expect(result[9]["predecessor"]).to.equal(result[8]["hash"]);
+        expect(parseInt(result[9]['level'])).to.greaterThan(parseInt(result[1]['level']));
+        expect(result[9]['predecessor']).to.equal(result[8]['hash']);
     });
-
+    
     it('retrieve a single an account', async () => {
-        var result = await TezosConseilClient.getAccount({url: ConseilV2URL, apiKey: ConseilV2APIKey}, 'alphanet', 'tz1XErrAm8vFBzu69UU74JUSbvsmvXiQBy6e');
+        var result = await TezosConseilClient.getAccount(conseilServer, 'mainnet', 'tz1bwSatfnvehPG8v5razwMJ7KzJXDBRtxwk');
 
         expect(result.length).to.equal(1);
         expect(result[0]['account_id']).to.equal('tz1XErrAm8vFBzu69UU74JUSbvsmvXiQBy6e', 'this may vary as the network changes');
@@ -71,12 +74,10 @@ describe('Tezos date interface test suite', () => {
         accountsquery = ConseilQueryBuilder.addPredicate(accountsquery, 'account_id', ConseilOperator.EQ, ['tz1aCy8b6Ls4Gz7m5SbANjtMPiH6dZr9nnS2'], true);
         accountsquery = ConseilQueryBuilder.addOrdering(accountsquery, 'block_level', ConseilSortDirection.DESC);
         accountsquery = ConseilQueryBuilder.setLimit(accountsquery, 300);
-        const result = await TezosConseilClient.getAccounts({url: ConseilV2URL, apiKey: ConseilV2APIKey}, 'alphanet', accountsquery);
+        const result = await TezosConseilClient.getAccounts(conseilServer, 'mainnet', accountsquery);
 
         expect(result.length).to.be.greaterThan(1, 'this may vary as the network changes');
         expect(result[0]['account_id']).to.equal('KT1SJdeXcP4KkVFF13SYEDFixunJBP7Nwmum', 'this may vary as the network changes');
-
-        //console.log(util.inspect(accountsquery, {showHidden: false, depth: 5, colors: false, maxArrayLength: 100}));
     });
 
     it('retrieve transactions for an account', async () => {
@@ -92,7 +93,7 @@ describe('Tezos date interface test suite', () => {
         target = ConseilQueryBuilder.addOrdering(target, 'block_level', ConseilSortDirection.DESC);
         target = ConseilQueryBuilder.setLimit(target, 300);
 
-        var result = await Promise.all([target, origin].map(q => TezosConseilClient.getOperations({url: ConseilV2URL, apiKey: ConseilV2APIKey}, 'alphanet', q)))
+        var result = await Promise.all([target, origin].map(q => TezosConseilClient.getOperations(conseilServer, 'mainnet', q)))
             .then(responses => responses.reduce((result, r) => { r.forEach(rr => result.push(rr)); return result; }));
 
         expect(result.length).to.be.greaterThan(10, 'this may vary as the network changes');
@@ -105,7 +106,7 @@ describe('Tezos date interface test suite', () => {
         operationFeesQuery = ConseilQueryBuilder.addOrdering(operationFeesQuery, 'block_level', ConseilSortDirection.DESC);
         operationFeesQuery = ConseilQueryBuilder.setLimit(operationFeesQuery, 1000);
 
-        const fees = await TezosConseilClient.getOperations({url: ConseilV2URL, apiKey: ConseilV2APIKey}, 'alphanet', operationFeesQuery);
+        const fees = await TezosConseilClient.getOperations(conseilServer, 'mainnet', operationFeesQuery);
         const sortedfees = fees.map(f => parseInt(f['fee'])).sort((a, b) => a - b);
 
         const lowAverageFee = sortedfees.slice(0, 300).reduce((s, c) => s + c) / 300;
@@ -123,7 +124,7 @@ describe('Tezos date interface test suite', () => {
         operationFeesQuery = ConseilQueryBuilder.addOrdering(operationFeesQuery, 'timestamp', ConseilSortDirection.DESC);
         operationFeesQuery = ConseilQueryBuilder.setLimit(operationFeesQuery, 1);
 
-        const fees = await TezosConseilClient.getTezosEntityData({url: ConseilV2URL, apiKey: ConseilV2APIKey}, 'alphanet', 'fees', operationFeesQuery);
+        const fees = await TezosConseilClient.getTezosEntityData(conseilServer, 'mainnet', 'fees', operationFeesQuery);
 
         expect(fees[0]['low']).to.lessThan(fees[0]['medium']);
         expect(fees[0]['medium']).to.lessThan(fees[0]['high']);
@@ -136,7 +137,7 @@ describe('Tezos date interface test suite', () => {
         operationFeesQuery = ConseilQueryBuilder.addOrdering(operationFeesQuery, 'timestamp', ConseilSortDirection.DESC);
         operationFeesQuery = ConseilQueryBuilder.setLimit(operationFeesQuery, 1);
 
-        const fees = await TezosConseilClient.getFeeStatistics({url: ConseilV2URL, apiKey: ConseilV2APIKey}, 'alphanet', OperationKindType.Transaction);
+        const fees = await TezosConseilClient.getFeeStatistics(conseilServer, 'mainnet', OperationKindType.Transaction);
 
         expect(fees[0]['low']).to.lessThan(fees[0]['medium']);
         expect(fees[0]['medium']).to.lessThan(fees[0]['high']);
